requirements.txt
flask==3.0.0
langchain==0.0.314
graphene>=3.1
gql
requests-toolbelt
openai
